Extract folder name functionality from getLibraryString() into getFolderName(). Some integration tests was in unit test folder
