+++ sourcecode/apis/contentauthor/app/Libraries/H5P/Storage/H5PCerpusStorage.php
@@ -498,7 +498,7 @@ class H5PCerpusStorage implements H5PFileStorage, H5PDownloadInterface, CerpusSt
     {
         /** @var H5PLibrary $library */
         $library = H5PLibrary::fromLibrary([$machineName, $majorVersion, $minorVersion])->latestVersion()->first();
-        $path = sprintf(ContentStorageSettings::UPGRADE_SCRIPT_PATH, $library->getLibraryString(true));
+        $path = sprintf(ContentStorageSettings::UPGRADE_SCRIPT_PATH, $library->getFolderName());
 
         return $this->filesystem->exists($path) ? "/$path" : null;
     }
