+++ sourcecode/apis/contentauthor/phpstan-baseline.neon
@@ -1672,6 +1672,11 @@ parameters:
 			path: tests/Integration/Libraries/H5P/CRUTest.php
 
 		-
+			message: "#^Parameter \\#1 \\$library of method App\\\\Libraries\\\\H5P\\\\Framework\\:\\:deleteLibrary\\(\\) expects stdClass, App\\\\H5PLibrary given\\.$#"
+			count: 1
+			path: tests/Integration/Libraries/H5P/FrameworkTest.php
+
+		-
 			message: "#^Access to an undefined property Illuminate\\\\Database\\\\Eloquent\\\\Model\\:\\:\\$id\\.$#"
 			count: 7
 			path: tests/Integration/Libraries/H5P/H5PCopyrightTest.php
@@ -1747,6 +1752,21 @@ parameters:
 			path: tests/Integration/Libraries/H5P/Package/VideoTest.php
 
 		-
+			message: "#^Access to an undefined property Illuminate\\\\Database\\\\Eloquent\\\\Model\\:\\:\\$auth_id\\.$#"
+			count: 2
+			path: tests/Integration/Libraries/H5P/ViewConfigTest.php
+
+		-
+			message: "#^Access to an undefined property Illuminate\\\\Database\\\\Eloquent\\\\Model\\:\\:\\$email\\.$#"
+			count: 2
+			path: tests/Integration/Libraries/H5P/ViewConfigTest.php
+
+		-
+			message: "#^Access to an undefined property Illuminate\\\\Database\\\\Eloquent\\\\Model\\:\\:\\$name\\.$#"
+			count: 1
+			path: tests/Integration/Libraries/H5P/ViewConfigTest.php
+
+		-
 			message: "#^Parameter \\$share of class App\\\\Libraries\\\\DataObjects\\\\ResourceMetadataDataObject constructor expects string\\|null, true given\\.$#"
 			count: 2
 			path: tests/Integration/Libraries/QuestionSetConverterTest.php
@@ -1815,23 +1835,3 @@ parameters:
 			message: "#^Method Tests\\\\TestCase\\:\\:setUpTraits\\(\\) should return array but return statement is missing\\.$#"
 			count: 1
 			path: tests/TestCase.php
-
-		-
-			message: "#^Parameter \\#1 \\$library of method App\\\\Libraries\\\\H5P\\\\Framework\\:\\:deleteLibrary\\(\\) expects stdClass, App\\\\H5PLibrary given\\.$#"
-			count: 1
-			path: tests/Unit/Libraries/H5P/FrameworkTest.php
-
-		-
-			message: "#^Access to an undefined property Illuminate\\\\Database\\\\Eloquent\\\\Model\\:\\:\\$auth_id\\.$#"
-			count: 2
-			path: tests/Unit/Libraries/H5P/ViewConfigTest.php
-
-		-
-			message: "#^Access to an undefined property Illuminate\\\\Database\\\\Eloquent\\\\Model\\:\\:\\$email\\.$#"
-			count: 2
-			path: tests/Unit/Libraries/H5P/ViewConfigTest.php
-
-		-
-			message: "#^Access to an undefined property Illuminate\\\\Database\\\\Eloquent\\\\Model\\:\\:\\$name\\.$#"
-			count: 1
-			path: tests/Unit/Libraries/H5P/ViewConfigTest.php
