+++ sourcecode/apis/contentauthor/tests/Integration/H5PContentTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Unit;
+namespace Tests\Integration;
 
 use App\H5PContent;
 use App\H5PLibrary;
