+++ sourcecode/apis/contentauthor/tests/Integration/Libraries/H5P/CRUTest.php
@@ -514,7 +514,7 @@ class CRUTest extends TestCase
                 '_token' => csrf_token(),
                 'title' => 'New resource',
                 'action' => 'create',
-                'library' => $library->getLibraryString(false, false),
+                'library' => $library->getLibraryString(false),
                 'parameters' => '{"params":{"simpleTest":"SimpleTest"},"metadata":{}}',
                 'license' => "PRIVATE",
                 'lti_message_type' => $this->faker->word,
@@ -537,7 +537,7 @@ class CRUTest extends TestCase
             ->put(route('h5p.update', $newContent->id), [
                 '_token' => csrf_token(),
                 'title' => $newContent->title,
-                'library' => $library->getLibraryString(false, false),
+                'library' => $library->getLibraryString(false),
                 'parameters' => '{"params":{"simpleTest":"SimpleTest"},"metadata":{}}',
                 'license' => "PRIVATE",
                 'lti_message_type' => $this->faker->word,
